# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/countries": {
      "get": {
        "operationId": "CountrySettingsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountrySettingResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all country settings",
        "tags": [
          "Admin - Country Settings"
        ]
      },
      "post": {
        "operationId": "CountrySettingsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountrySettingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create country setting",
        "tags": [
          "Admin - Country Settings"
        ]
      }
    },
    "/admin/countries/{id}": {
      "patch": {
        "operationId": "CountrySettingsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountrySettingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update country setting",
        "tags": [
          "Admin - Country Settings"
        ]
      },
      "delete": {
        "operationId": "CountrySettingsController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete country setting",
        "tags": [
          "Admin - Country Settings"
        ]
      }
    },
...
    "/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: get product by ID",
        "tags": [
          "Admin - Products"
        ]
      },
      "patch": {
        "operationId": "AdminProductsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: update product",
        "tags": [
          "Admin - Products"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_archive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: archive product",
        "tags": [
          "Admin - Products"
        ]
      }
    },
    "/admin/products/{productId}/images": {
      "get": {
        "operationId": "ProductImagesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductImageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: get product images",
        "tags": [
          "Admin - Product Images"
        ]
      },
      "post": {
        "operationId": "ProductImagesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: add product image",
        "tags": [
          "Admin - Product Images"
        ]
      }
    },
    "/admin/products/{productId}/images/{imageId}": {
      "patch": {
        "operationId": "ProductImagesController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: update product image",
        "tags": [
          "Admin - Product Images"
        ]
      },
      "delete": {
        "operationId": "ProductImagesController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: delete product image",
        "tags": [
          "Admin - Product Images"
        ]
      }
    },
    "/admin/products/{productId}/images/{imageId}/default": {
      "patch": {
        "operationId": "ProductImagesController_setDefault",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: set product image as default",
        "tags": [
          "Admin - Product Images"
        ]
      }
    },
...
    "/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "parameters": [
          {
            "name": "x-guest-id",
            "in": "header",
            "description": "Required only for guest cart requests. Not required when Authorization bearer token is valid.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user or guest cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "parameters": [
          {
            "name": "x-guest-id",
            "in": "header",
            "description": "Required only for guest cart requests. Not required when Authorization bearer token is valid.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add item to cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/items/{itemId}": {
      "patch": {
        "operationId": "CartController_updateItem",
        "parameters": [
          {
            "name": "x-guest-id",
            "in": "header",
            "description": "Required only for guest cart requests. Not required when Authorization bearer token is valid.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update cart item quantity",
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "x-guest-id",
            "in": "header",
            "description": "Required only for guest cart requests. Not required when Authorization bearer token is valid.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove item from cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/clear": {
      "delete": {
        "operationId": "CartController_clearCart",
        "parameters": [
          {
            "name": "x-guest-id",
            "in": "header",
            "description": "Required only for guest cart requests. Not required when Authorization bearer token is valid.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Clear cart",
        "tags": [
          "Cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zemlo API",
    "description": "Zemlo e-commerce backend API contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "CountrySettingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iso2": {
            "type": "string"
          },
          "iso3": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "phoneCode": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "allowWebsiteAccess": {
            "type": "boolean"
          },
          "allowCheckout": {
            "type": "boolean"
          },
          "allowShipping": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "iso2",
          "isActive",
          "allowWebsiteAccess",
          "allowCheckout",
          "allowShipping",
          "isDefault",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "AdminProductCategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AdminProductBrandResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AdminProductImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "number"
          },
          "isDefault": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "url",
          "position",
          "isDefault",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminProductVariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "nullable": true
          },
          "compareAtPrice": {
            "type": "number",
            "nullable": true
          },
          "stock": {
            "type": "number"
          },
          "trackInventory": {
            "type": "boolean"
          },
          "allowBackorder": {
            "type": "boolean"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "type": "object"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "name",
          "stock",
          "trackInventory",
          "allowBackorder",
          "options",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number"
          },
          "compareAtPrice": {
            "type": "number",
            "nullable": true
          },
          "costPrice": {
            "type": "number",
            "nullable": true
          },
          "stock": {
            "type": "number"
          },
          "trackInventory": {
            "type": "boolean"
          },
          "allowBackorder": {
            "type": "boolean"
          },
          "hasVariants": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "INACTIVE",
              "ARCHIVED"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          },
          "categoryId": {
            "type": "string"
          },
          "brandId": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "nullable": true
          },
          "length": {
            "type": "number",
            "nullable": true
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metaTitle": {
            "type": "string",
            "nullable": true
          },
          "metaDescription": {
            "type": "string",
            "nullable": true
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/AdminProductCategoryResponseDto"
          },
          "brand": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminProductBrandResponseDto"
              }
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminProductImageResponseDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminProductVariantResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "stock",
          "trackInventory",
          "allowBackorder",
          "hasVariants",
          "status",
          "isFeatured",
          "categoryId",
          "keywords",
          "category",
          "images",
          "variants",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "number"
          },
          "isDefault": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "url",
          "position",
          "isDefault",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CartProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number"
          },
          "compareAtPrice": {
            "type": "number",
            "nullable": true
          },
          "stock": {
            "type": "number"
          },
          "trackInventory": {
            "type": "boolean"
          },
          "allowBackorder": {
            "type": "boolean"
          },
          "hasVariants": {
            "type": "boolean"
          },
          "category": {
            "$ref": "#/components/schemas/CartCategoryResponseDto"
          },
          "brand": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CartBrandResponseDto"
              }
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartImageResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "stock",
          "trackInventory",
          "allowBackorder",
          "hasVariants",
          "category",
          "images"
        ]
      },
      "CartVariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "nullable": true
          },
          "compareAtPrice": {
            "type": "number",
            "nullable": true
          },
          "stock": {
            "type": "number"
          },
          "trackInventory": {
            "type": "boolean"
          },
          "allowBackorder": {
            "type": "boolean"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "stock",
          "trackInventory",
          "allowBackorder",
          "options"
        ]
      },
      "CartItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "cartId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "variantId": {
            "type": "string",
            "nullable": true
          },
          "variantKey": {
            "type": "string"
          },
          "addedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "unitPrice": {
            "type": "number"
          },
          "lineTotal": {
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/CartProductResponseDto"
          },
          "variant": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CartVariantResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "quantity",
          "cartId",
          "productId",
          "variantKey",
          "addedAt",
          "updatedAt",
          "unitPrice",
          "lineTotal",
          "product"
        ]
      },
      "CartResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "guestId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemResponseDto"
            }
          },
          "totalItems": {
            "type": "number"
          },
          "totalQuantity": {
            "type": "number"
          },
          "subtotal": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "items",
          "totalItems",
          "totalQuantity",
          "subtotal"
        ]
      },
...
    }
  }
}